+++ CHANGELOG.md
@@ -1,12 +1,22 @@
 Change Log
 ==========
 
+Version 1.0.3 *(2019-8-31)*
+----------------------------
+
+* fix #2，新增四周阴影功能
+  * ad_shadowSize 设置阴影宽度
+  * ad_shadowColor 设置阴影颜色
+  * ArrowDrawable 新增 setShadowSizeDp 和 setShadowSizeRes 方法设置阴影宽度 
+  * ArrowDrawable 新增 setShadowColor 和 setShadowColorRes 方法设置阴影颜色
+* 之前的背景色属性名 ad_background 改为 ad_bgColor
+
 Version 1.0.1 *(2019-8-22)*
 ----------------------------
 
-- TransformersTip 新增 setBackgroundDimEnabled 方法，设置是否允许浮窗的背景变暗
+* TransformersTip 新增 setBackgroundDimEnabled 方法，设置是否允许浮窗的背景变暗
 
 Version 1.0.0 *(2019-8-21)*
 ----------------------------
 
-Initial release.
+* Initial release